Frontend
- added multiple markers for tooltip in highcharts
- added are_you_sure modal following delete run button
